+++ CHANGELOG.md
@@ -8,7 +8,7 @@ Versions follow [CalVer](https://calver.org).
 
 - The Project now uses CalVer to define versions and releases.
 - Add routines for creating spatial and temporal meshes/grids.
-- Add class `BEPSSolver` to find a 1D Hamiltonian ground-state using the  _Backward Euler
+- Add class `BEPSSolver` to find a 1D Hamiltonian ground-state using the _Backward Euler
   Pseudo-Spectral_ (BEPS) method.
 - Define a classes and protocols to define one-dimensional Hamiltonians.
   - The class `HTHamiltonian` represents a 1D Bose gas within a harmonic trap